Add the beginnings of a Dart tooling MCP server (#9)
This service demonstrates our ability to connect to the dart tooling daemon to perform tasks, as well as the vm service connections we can access through that.

Currently screenshots work for flutter apps, but we can easily plug in more.

In the future we likely are going to want to make the DTD connection more lazy - so that we don't require it on startup, but instead ask for it as a part of the tool methods that require it. This should result in a more friendly user experience for setting up the tool (assuming the clients prompt them for it).

TODO: tests, which will likely require some significant setup (flutter SDK + active DTD connection), hopefully we can leverage some existing devtools testing functionality for this.