move ServerImplementation and instructions to a final field (#39)
There was a bit of a discrepancy between the MCPClient and MCPServer interfaces here, the client had a final field you had to initialize in the constructor, and the server made you override abstract getters.

I opted for the MCPClient pattern here, but I don't think there is any strong reason to choose one over the other, it just needs to be consistent.