Add basic analysis support (#43)
Requires clients to implement "Roots", and creates one analysis context per root.

Then, requests to analyze groups of files can be issued via the "analyze files" command, which requires a root for each group of files.

We also run a file watcher, which is a bit weird right now because we could be a second or so behind if using a polling watcher. We might want to try and do some sort of "file overlay" situation for other scenarios, passing the file contents to analyze directly.