Add resource template support (#68)
See https://modelcontextprotocol.io/specification/2024-11-05/server/resources#resource-templates

The support is pretty low level right now, and we do not provide any assistance for matching the provided templates or extracting out the values.

We likely should add that support at some point later on, but we should be able to do so in a non-breaking manner. We will simply stop invoking the `handlers` for URIs that can't possibly match the template.