Break pub tools out into their own support mixin (#82)
This refactors the pub support to be part of its own support mixin instead of being part of the Dart CLI tools mixin. This is because once https://github.com/dart-lang/ai/issues/81 is resolved, this tool won't necessarily be using the Dart CLI or the Flutter CLI, but may use either.

This PR pulls out a utility method `runCommandInRoots` from the Dart CLI support mixin for reuse.