Add runtime errors resource and tool to clear errors. (#94)
This ended up being a fair bit more complicated than I wanted, as it also refactors things and moves us towards a direction where we can support multiple connected apps (which was a natural extension of creating the resource URI, to include the debug session ID in it).

Functionally, this:

- Starts collecting errors as soon as a debug session is connected, and creates a runtime errors resource for each session.
- Allows us to fire change notifications for when new errors come in.
- Also added a `clear_runtime_errors` tool as well as an argument to the `hotReload` tool to clear the errors prior to the hot reload.
- Drops the support for `since` in favor of just clearing the errors.

Eventually we might want to support pagination of the resource instead (it is already modeled as many parts), I filed a feature request on the MCP repo for this, but it likely won't land for some time if at all.