handle failing to listen to vm service streams during startup (#124)
This should resolve the flakyness in tests.

Other VM service interactions already handle this better because they are scoped to an RPC request, and will just return an appropriate error for that request. But we set up these error listeners out of band and so the exceptions are not handled if we fail to listen to the streams due to the connection being closed.