Add RootsFallbackSupport mixin (#132)
This mixin can be used to conditionally add `addRoots` and `removeRoots` tools, when the client doesn't support roots.

It overrides all the `RootsTrackingSupport` functionality as well as the `listRoots` method, to use the custom behavior instead.

Also adds a flag to force this fallback mode to the server, which is required for some clients who claim to support roots but do not actually (like cursor).

Closes https://github.com/dart-lang/ai/issues/111