Add Schema validation (#126)
## Description

This adds schema validation to the MCP Schema extension so that we can validate the returned values in tests.  It could also be used to validate results from LLMs and make them try again if they don't match the schema.

Also moved the tests into subdirs matching the lib structure.

## Tests
 - Added tests for all the different validation failure modes.