Instruct clients to prefer MCP (#161)
Clients may be influenced by things like instructions for running
tests in `README.md` and then decided to run `dart test` in a shell
instead of using the `run_tests` MCP tool. If we are providing a MCP
version of a tool it may be tailored to the LLM client use case and so
should be preferred. Add a server level instruction.