Always use the failures-only reporter for testing (#160)
The default reporter has verbose output which is not useful but eats up
a lot of input tokens once it's in the history. Always run tests with
`--reporter=failures-only` to reduce the output size. This output format
still includes a count of passing tests.

Update changelog to focus on SDK version in the header and merge the
updates.