Add flutter driver tool (#223)
Towards https://github.com/dart-lang/ai/issues/222 (could close it, but probably we should do some more work to validate it works well).

Adds the flutter driver tool, with a curated set of available commands. I decided to expose this as just a single tool, which mirrors exactly the flutter driver extension method.

The schema for this is pretty weird and complex - but in general Gemini seems to do OK with it.

Note that I only tested the tap/get_text commands, I think that is likely sufficient (really we just need to test we are forwarding commands through, we aren't doing anything special).

Also updates the workflow client to handle enums and use updated model names.